requirements.txt
streamlit>=1.38,<2.0
tensorflow==2.12.0
numpy>=1.22,<1.24
pillow>=10.0,<11.0
pandas>=2.0,<3.0
matplotlib>=3.7,<4.0
h5py>=3.0,<4.0
